+++ PSModule/IntuneDocumentation/Internal/Set-AzureADAppPermission.ps1
@@ -1,0 +1,43 @@
+Function Set-AzureADAppPermission
+{
+    [CmdletBinding(SupportsShouldProcess, ConfirmImpact='Medium')]
+    param
+    (
+        [string] $targetServicePrincipalName,
+        $appPermissionsRequired,
+        $childApp,
+        $spForApp
+    )
+
+    $targetSp = Get-AzureADServicePrincipal -Filter "DisplayName eq '$($targetServicePrincipalName)'"
+
+    # Iterate Permissions array
+    Write-Output -InputObject ('Retrieve Role Assignments objects')
+    $RoleAssignments = @()
+    Foreach ($AppPermission in $appPermissionsRequired) {
+        $RoleAssignment = $targetSp.AppRoles | Where-Object { $_.Value -eq $AppPermission}
+        $RoleAssignments += $RoleAssignment
+    }
+
+    $ResourceAccessObjects = New-Object 'System.Collections.Generic.List[Microsoft.Open.AzureAD.Model.ResourceAccess]'
+    foreach ($RoleAssignment in $RoleAssignments) {
+        $resourceAccess = New-Object -TypeName "Microsoft.Open.AzureAD.Model.ResourceAccess"
+        $resourceAccess.Id = $RoleAssignment.Id
+        $resourceAccess.Type = 'Role'
+        $ResourceAccessObjects.Add($resourceAccess)
+    }
+    $requiredResourceAccess = New-Object -TypeName "Microsoft.Open.AzureAD.Model.RequiredResourceAccess"
+    $requiredResourceAccess.ResourceAppId = $targetSp.AppId
+    $requiredResourceAccess.ResourceAccess = $ResourceAccessObjects
+
+    # set the required resource access
+    Set-AzureADApplication -ObjectId $childApp.ObjectId -RequiredResourceAccess $requiredResourceAccess
+    Start-Sleep -s 1
+
+    # grant the required resource access
+    foreach ($RoleAssignment in $RoleAssignments) {
+        Write-Output -InputObject ('Granting admin consent for App Role: {0}' -f $($RoleAssignment.Value))
+        New-AzureADServiceAppRoleAssignment -ObjectId $spForApp.ObjectId -Id $RoleAssignment.Id -PrincipalId $spForApp.ObjectId -ResourceId $targetSp.ObjectId
+        Start-Sleep -s 1
+    }
+}